+++ inst/system/create_prevfit.R
@@ -1,4 +1,7 @@
-library("adjclust") #primitive version 0.3.0
+if (packageVersion("adjclust") > "0.3.0") {
+    stop("Please install a version of 'adjclust' <= 0.3.0 to run this script")
+}
+library("adjclust")
 library("snpStats")
 
 data("ld.example", package="snpStats")
@@ -10,6 +13,6 @@ ceph.1mb[4,286]@.Data[1,1] <- as.raw(3) #to avoid NaNs
 ld.ceph <- ld(ceph.1mb, stats="R.squared", depth=h)
 ld.ceph <- round(ld.ceph, digits=10)
 
-prevfit <- adjclust:::adjClustBand_heap(ld.ceph@x, p, h, blMin=1)
+res_adjclust_0.3.0 <- adjclust:::adjClustBand_heap(ld.ceph@x, p, h, blMin=1)
 
-save(prevfit, file="data/prevfit.rda", compress="xz")
+save(res_adjclust_0.3.0, file="data/res_adjclust_0.3.0.rda", compress="xz")