+++ NEWS.md
@@ -1,5 +1,6 @@
-# Version 0.4.2 [2017-10-01]
+# Version 0.4.2 [2017-10-03]
 
+* Improved handling of default value for 'h' in 'adjclust' for 'dist' objects
 * Renamed 'prevfit' into the more explicit 'res_adjclust_0.3.0'
 * Dropped 'simmatrix' toy data set (now generated on the fly in tests)